Autoloading issues
Namespace PSR-0 commit
Updated composers.json
Regroup multiple commit
