+++ lib/Payplug/Core/HttpClient.php
@@ -2,122 +2,6 @@
 namespace Payplug\Core;
 
 /**
- * Generified HttpRequest so that it can easily be mocked
- */
-interface IHttpRequest
-{
-    /**
-     * Simple wrapper for curl_setopt
-     *
-     * @link http://php.net/manual/en/function.curl-setopt.php
-     */
-    function setopt($option, $value);
-
-    /**
-     * Simple wrapper for curl_exec
-     *
-     * @link http://php.net/manual/en/function.curl-exec.php
-     */
-    function exec();
-
-    /**
-     * Simple wrapper for curl_getinfo
-     *
-     * @link http://php.net/manual/en/function.curl-getinfo.php
-     */
-    function getinfo($option);
-
-    /**
-     * Simple wrapper for curl_close
-     *
-     * @link http://php.net/manual/en/function.curl-close.php
-     */
-    function close();
-
-    /**
-     * Simple wrapper for curl_error
-     *
-     * @link http://php.net/manual/en/function.curl-close.php
-     */
-    function error();
-
-    /**
-     * Simple wrapper for curl_errno
-     *
-     * @link http://php.net/manual/en/function.curl-close.php
-     */
-    function errno();
-}
-
-/**
- * A Curl implementation of a IHttpRequest.
- */
-class CurlRequest implements IHttpRequest
-{
-    /**
-     * @var resource the curl object
-     */
-    private $_curl;
-
-    /**
-     * \Payplug\CurlRequest constructor
-     * Initializes a curl request
-     */
-    public function __construct()
-    {
-        $this->_curl = curl_init();
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function setopt($option, $value)
-    {
-        return curl_setopt($this->_curl, $option, $value);
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function getinfo($option)
-    {
-        return curl_getinfo($this->_curl, $option);
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function exec()
-    {
-        return curl_exec($this->_curl);
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function close()
-    {
-        curl_close($this->_curl);
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function error()
-    {
-        return curl_error($this->_curl);
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function errno()
-    {
-        return curl_errno($this->_curl);
-    }
-}
-
-/**
  * An authenticated HTTP client for PayPlug's API.
  *
  * @note This requires PHP's curl extension.
