+++ lib/Payplug/Exception/BadRequestException.php
@@ -1,0 +1,9 @@
+<?php
+namespace Payplug\Exception;
+
+/**
+ * 400 Bad Request
+ */
+class BadRequestException extends HttpException
+{
+}
