+++ lib/Payplug/Exception/ConfigurationException.php
@@ -1,0 +1,9 @@
+<?php
+namespace Payplug\Exception;
+
+/**
+ * Thrown when there was a problem with a Authentication.
+ */
+class ConfigurationException extends \Payplug\Exception\PayPlugException
+{
+}
