+++ lib/Payplug/Exception/ForbiddenException.php
@@ -1,0 +1,9 @@
+<?php
+namespace Payplug\Exception;
+
+/**
+ * 403 Forbidden
+ */
+class ForbiddenException extends HttpException
+{
+}
