+++ lib/Payplug/Exception/HttpException.php
@@ -52,52 +52,3 @@ class HttpException extends \Payplug\Exception\PayPlugException
         return json_decode($this->_httpResponse, true);
     }
 }
-
-/**
- * 400 Bad Request
- */
-class BadRequestException extends HttpException
-{
-}
-
-/**
- * 401 Unauthorized
- */
-class UnauthorizedException extends HttpException
-{
-}
-
-/**
- * 403 Forbidden
- */
-class ForbiddenException extends HttpException
-{
-}
-
-/**
- * 404 Not Found
- */
-class NotFoundException extends HttpException
-{
-}
-
-/**
- * 405 Not Allowed
- */
-class NotAllowedException extends HttpException
-{
-}
-
-/**
- * 5XX server errors
- */
-class PayPlugServerException extends HttpException
-{
-}
-
-/**
- * Thrown when we expected the API to have a specific format, and we got something else.
- */
-class UnexpectedAPIResponseException extends HttpException
-{
-}
