+++ lib/Payplug/Exception/NotAllowedException.php
@@ -1,0 +1,9 @@
+<?php
+namespace Payplug\Exception;
+
+/**
+ * 405 Not Allowed
+ */
+class NotAllowedException extends HttpException
+{
+}
