+++ lib/Payplug/Exception/NotFoundException.php
@@ -1,0 +1,9 @@
+<?php
+namespace Payplug\Exception;
+
+/**
+ * 404 Not Found
+ */
+class NotFoundException extends HttpException
+{
+}
