+++ lib/Payplug/Exception/PHPVersionException.php
@@ -1,0 +1,9 @@
+<?php
+namespace Payplug\Exception;
+
+/**
+ * Wrong PHP version exception.
+ */
+class PHPVersionException extends DependencyException
+{
+}
