+++ lib/Payplug/Exception/PayPlugServerException.php
@@ -1,0 +1,9 @@
+<?php
+namespace Payplug\Exception;
+
+/**
+ * 5XX server errors
+ */
+class PayPlugServerException extends HttpException
+{
+}
