+++ lib/Payplug/Exception/UnauthorizedException.php
@@ -1,0 +1,9 @@
+<?php
+namespace Payplug\Exception;
+
+/**
+ * 401 Unauthorized
+ */
+class UnauthorizedException extends HttpException
+{
+}
