+++ lib/Payplug/Exception/UnexpectedAPIResponseException.php
@@ -1,0 +1,9 @@
+<?php
+namespace Payplug\Exception;
+
+/**
+ * Thrown when we expected the API to have a specific format, and we got something else.
+ */
+class UnexpectedAPIResponseException extends HttpException
+{
+}
