+++ lib/Payplug/Resource/APIResource.php
@@ -2,21 +2,6 @@
 namespace Payplug\Resource;
 
 /**
- * Interface designed to force resources to implement at least one factory.
- */
-interface IAPIResourceFactory
-{
-    /**
-     * The factory method that constructs the API resource.
-     *
-     * @param   array   $attributes the default attributes.
-     *
-     * @return  APIResource The new resource.
-     */
-    static function fromAttributes(array $attributes);
-}
-
-/**
  * A simple API resource
  */
 abstract class APIResource implements IAPIResourceFactory
